+++ src/Credfeto.ChangeLog.Tests/ChangeLogUpdaterAddEntryTests.cs
@@ -1,4 +1,5 @@
-﻿using Xunit;
+﻿using Credfeto.ChangeLog.Exceptions;
+using Xunit;
 
 namespace Credfeto.ChangeLog.Tests
 {
@@ -178,6 +179,34 @@ Releases that have at least been deployed to staging, BUT NOT necessarily releas
 ## [0.0.0] - Project created";
 
             Assert.Equal(expected: expected, actual: result);
+        }
+
+        [Fact]
+        public void AddToExistingChangelogForSectionThatDoesNotExistFails()
+        {
+            const string existing = @"# Changelog
+All notable changes to this project will be documented in this file.
+
+<!--
+Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
+-->
+
+## [Unreleased]
+### Fixed
+### Changed
+### Removed
+### Deployment Changes
+
+<!--
+Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch
+-->
+## [0.0.1] - 2020-12-29
+### Added
+- Added a new entry
+
+## [0.0.0] - Project created";
+
+            Assert.Throws<InvalidChangeLogException>(() => ChangeLogUpdater.AddEntry(changeLog: existing, type: "Added", message: "Another entry"));
         }
     }
 }